+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 **Added**:
 
 - Added method `ethereum_address` to `Payment`
+- Added method `get_current_locale` for `BaseProvider`
+
+
+**Updated**:
+
+- Updated method `version` for supporting pre-releases
 
 ---